+++ CHANGELOG.md
@@ -31,6 +31,16 @@ and this project adheres to [Semantic Versioning][Semantic Versioning].
 - Function that writes to excel
 - Buttons to add to ribbon: 100% zoom, zoom One Page, Advanced Document Properties, Sharepoint Properties, Accept all changes and stop tracking, Print preview, Open MS Excel, paste plain format
 
+## [0.0.1] - 2025-04-17
+
+### Added
+
+- link between doc properties and the SaveFileView form auto fills when user presses search btn
+
+### Changed
+
+- divided the `ConfigHelper.vb` into `PathHelper` and `ListHelper`
+
 ## [0.0.1] - 2025-04-16
 
 ### Added